:card_file_box: (db): 更新数据库版本
将数据库版本从 2 更新到 3，并添加迁移操作。

+++ server/src/main/java/org/ezbook/server/db/model/BillInfoModel.kt
@@ -98,7 +98,7 @@ class BillInfoModel {
     /**
      * 分组id，这个id是指将短时间内捕获到的同等金额进行合并的分组id
      */
-    var groupId: Int = -1
+    var groupId: Long = -1
 
     /**
      * 数据渠道，这里指的是更具体的渠道，例如【建设银行】微信公众号，用户【xxxx】这种
